+++ login-workflow/src/routing/AuthNavigationContainer.tsx
@@ -9,7 +9,7 @@ import { ForgotPassword } from '../screens/ForgotPassword';
 import { ResetPassword } from '../screens/ResetPassword';
 import { InviteRegistrationPager } from '../screens/InviteRegistrationPager';
 import { SelfRegistrationPager } from '../screens/SelfRegistrationPager';
-import { ContactSupport } from '../screens/subScreens/ContactSupport';
+import { ContactSupportSubscreen } from '../screens/subScreens/ContactSupportSubscreen';
 import { Login } from '../screens/Login';
 
 export type RouteConfig = {
@@ -146,7 +146,7 @@ export const AuthNavigationContainer: React.FC<
                                     <GuestGuard>
                                         {showContactSupport ? (
                                             <BrandedCardContainer>
-                                                <ContactSupport />
+                                                <ContactSupportSubscreen />
                                             </BrandedCardContainer>
                                         ) : (
                                             RedirectToLogin
