--- login-workflow/src/screens/subScreens/ContactSupport/index.ts
@@ -1,1 +1,0 @@
-export * from './ContactSupport';
